exception delete

+++ src/main/java/com/system/dscommerce/services/ProductService.java
@@ -1,14 +1,17 @@
 package com.system.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.system.dscommerce.dtos.ProductDTO;
 import com.system.dscommerce.entities.Product;
 import com.system.dscommerce.repositories.ProductRepository;
+import com.system.dscommerce.services.exceptions.DatabaseExceptionService;
 import com.system.dscommerce.services.exceptions.NotFoundExceptionService;
 
 import jakarta.persistence.EntityNotFoundException;
@@ -51,12 +54,14 @@ public class ProductService {
         }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id))
+            throw new NotFoundExceptionService("Produto não encontrado para ser deletado");
         try {
             repository.deleteById(id);
-        } catch (IllegalArgumentException e) {
-            throw new NotFoundExceptionService("Produto não encontrado para ser deletado");
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseExceptionService("Produto não pode ser deletado");
         }
     }
